+++ source/Zoho.Api/Contract/IZsSubscriptionApi.cs
@@ -14,7 +14,8 @@
         Task<ZsSubscription> GetAsync(string authToken, string organizationId, string id);
 
         Task<ZsSubscriptions> GetAllAsync();
-        Task<ZsSubscriptions> GetAllAsync(string authToken, string organizationId);
+        Task<ZsSubscriptions> GetAllAsync(ZsCustomer customer);
+        Task<ZsSubscriptions> GetAllAsync(string authToken, string organizationId, ZsCustomer customer = null);
 
         Task<ZsSubscription> CreateAsync(ZsSubscriptionCreate createInput);
         Task<ZsSubscription> CreateAsync(string authToken, string organizationId, ZsSubscriptionCreate createInput);
